+++ code/cgame/cg_particles.c
@@ -722,8 +722,8 @@ void CG_AddParticleToScene( cparticle_t *p, vec3_t org, float alpha ) {
 			height = p->endheight;
 		}
 
-		sinR = height * sin( DEG2RAD( p->roll ) ) * sqrt( 2 );
-		cosR = width * cos( DEG2RAD( p->roll ) ) * sqrt( 2 );
+		sinR = height * sinf( DEG2RAD( p->roll ) ) * sqrt( 2 );
+		cosR = width * cosf( DEG2RAD( p->roll ) ) * sqrt( 2 );
 
 		VectorCopy( org, verts[0].xyz );
 		verts[0].xyz[0] -= sinR;
